py: refactoring: pyTypeFromIdentifier now lives in naming-utils.ts
